Add more lua teplates
