+++ cli/connect/templates/eval_func_body.lua
@@ -1,0 +1,1 @@
+return require('console').eval(...)
