+++ cli/connect/templates/title_func_body.lua
@@ -1,0 +1,10 @@
+
+local ok, api_topology = pcall(require, 'cartridge.lua-api.topology')
+if not ok then
+	return ''
+end
+
+local self = api_topology.get_self()
+if self.app_name == nil or self.instance_name == nil then
+	return ''
+end
