+++ cli/connector/eval_plain_text.go
@@ -12,6 +12,7 @@ import (
 	"github.com/vmihailenco/msgpack/v5"
 	lua "github.com/yuin/gopher-lua"
 
+	"github.com/tarantool/cartridge-cli/cli/codegen/static"
 	"github.com/tarantool/cartridge-cli/cli/templates"
 	"gopkg.in/yaml.v2"
 )
@@ -40,6 +41,11 @@ type PlainTextEvalRes struct {
 // Function should return `interface{}`, `string` (res, err)
 // to be correctly processed.
 func callPlainTextConn(conn net.Conn, funcName string, args []interface{}, opts EvalPlainTextOpts) ([]interface{}, error) {
+	callFuncTmpl, err := static.GetStaticFileContent(ConnectorLuaTemplateFS, "call_func.lua")
+	if err != nil {
+		return nil, fmt.Errorf("Failed to get static file content: %s", err)
+	}
+
 	evalFunc, err := templates.GetTemplatedStr(&callFuncTmpl, map[string]string{
 		"FunctionName": funcName,
 	})
@@ -55,6 +61,11 @@ func callPlainTextConn(conn net.Conn, funcName string, args []interface{}, opts 
 // Function should return `interface{}`, `string` (res, err)
 // to be correctly processed.
 func evalPlainTextConn(conn net.Conn, funcBody string, args []interface{}, opts EvalPlainTextOpts) ([]interface{}, error) {
+	evalFuncTmpl, err := static.GetStaticFileContent(ConnectorLuaTemplateFS, "eval_func.lua")
+	if err != nil {
+		return nil, err
+	}
+
 	if err := formatAndSendEvalFunc(conn, funcBody, args, evalFuncTmpl); err != nil {
 		return nil, err
 	}
@@ -85,6 +96,11 @@ func formatAndSendEvalFunc(conn net.Conn, funcBody string, args []interface{}, e
 	argsEncoded, err := msgpack.Marshal(args)
 	if err != nil {
 		return fmt.Errorf("Failed to encode args: %s", err)
+	}
+
+	evalFuncTmpl, err = static.GetStaticFileContent(ConnectorLuaTemplateFS, "eval_func.lua")
+	if err != nil {
+		return fmt.Errorf("Failed to get static file content: %s", err)
 	}
 
 	evalFunc, err := templates.GetTemplatedStr(&evalFuncTmpl, map[string]string{
@@ -390,28 +406,3 @@ func getPlainTextEvalResLua(resBytes []byte) (string, error) {
 
 	return encodedData, nil
 }
-
-var (
-	callFuncTmpl = `return {{ .FunctionName }}(...)`
-)
-
-const (
-	evalFuncTmpl = `
-local function func(...)
-	{{ .FunctionBody }}
-end
-local args = require('msgpack').decode(string.fromhex('{{ .ArgsEncoded }}'))
-
-local ret = {
-	load(
-		'local func, args = ... return func(unpack(args))',
-		'@eval'
-	)(func, args)
-}
-return {
-	data_enc = require('digest').base64_encode(
-		require('msgpack').encode(ret)
-	)
-}
-`
-)
