+++ cli/connector/templates/call_func.lua
@@ -1,0 +1,1 @@
+return {{ .FunctionName }}(...)
