+++ cli/repair/patch.go
@@ -6,6 +6,7 @@ import (
 	"path/filepath"
 	"strings"
 
+	"github.com/tarantool/cartridge-cli/cli/codegen/static"
 	"github.com/tarantool/cartridge-cli/cli/connector"
 
 	"github.com/tarantool/cartridge-cli/cli/common"
@@ -96,6 +97,13 @@ func reloadConf(topologyConfPath string, instanceName string, ctx *context.Ctx) 
 
 	// eval
 	confPath := filepath.Dir(topologyConfPath)
+
+	reloadClusterwideConfigFuncBody, err := static.GetStaticFileContent(RepairLuaTemplateFS,
+		"reload_clusterwide_config_func_body.lua")
+	if err != nil {
+		return resMessages, fmt.Errorf("Failed to get static file content: %s", err)
+	}
+
 	req := connector.EvalReq(reloadClusterwideConfigFuncBody, confPath, confapplierWishStateTimeout)
 
 	var results []bool
@@ -115,46 +123,3 @@ func reloadConf(topologyConfPath string, instanceName string, ctx *context.Ctx) 
 
 	return resMessages, nil
 }
-
-const (
-	reloadClusterwideConfigFuncBody = `
-local ClusterwideConfig = require('cartridge.clusterwide-config')
-local confapplier = require('cartridge.confapplier')
-
-local conf_path, wish_state_timeout = ...
-
-local cfg, err = ClusterwideConfig.load(conf_path)
-assert(err == nil, string.format('Failed to load new config: %s', err))
-
-local current_uuid = box.info().uuid
-if cfg:get_readonly().topology.servers[current_uuid] == nil then
-	return false
-end
-
-local roles_configured_state = 'RolesConfigured'
-local connecting_fullmesh_state = 'ConnectingFullmesh'
-
-local state = confapplier.wish_state(roles_configured_state, wish_state_timeout)
-
-if state == connecting_fullmesh_state then
-	error(string.format(
-		'Failed to reach %s config state. Stuck in %s. ' ..
-			'Call "box.cfg({replication_connect_quorum = 0})" in instance console and try again',
-		roles_configured_state, state
-	))
-end
-
-if state ~= roles_configured_state then
-	error(string.format(
-		'Failed to reach %s config state. Stuck in %s',
-		roles_configured_state, state
-	))
-end
-
-cfg:lock()
-local ok, err = confapplier.apply_config(cfg)
-assert(ok, string.format('Failed to apply new config: %s', err))
-
-return true
-`
-)
