+++ cli/replicasets/membership.go
@@ -3,6 +3,7 @@ package replicasets
 import (
 	"fmt"
 
+	"github.com/tarantool/cartridge-cli/cli/codegen/static"
 	"github.com/tarantool/cartridge-cli/cli/common"
 	"github.com/tarantool/cartridge-cli/cli/connector"
 	"github.com/vmihailenco/msgpack/v5"
@@ -35,6 +36,11 @@ func connectToMembership(conn *connector.Conn, runningInstancesNames []string, i
 		urisToProbe = append(urisToProbe, instanceConf.URI)
 	}
 
+	probeInstancesBody, err := static.GetStaticFileContent(ReplicasetsLuaTemplateFS, "probe_instances_body.lua")
+	if err != nil {
+		return fmt.Errorf("Failed to get static file content: %s", err)
+	}
+
 	if _, err := conn.Exec(connector.EvalReq(probeInstancesBody, urisToProbe)); err != nil {
 		return fmt.Errorf("Failed to probe all instances mentioned in replica sets: %s", err)
 	}
@@ -44,6 +50,12 @@ func connectToMembership(conn *connector.Conn, runningInstancesNames []string, i
 
 func getMembershipInstancesFromConn(conn *connector.Conn) (*MembershipInstances, error) {
 	var membershipInstancesSlice []*MembershipInstance
+
+	getMembershipInstancesBody, err := static.GetStaticFileContent(ReplicasetsLuaTemplateFS,
+		"membership_instances_body.lua")
+	if err != nil {
+		return nil, fmt.Errorf("Failed to get static file content: %s", err)
+	}
 
 	req := connector.EvalReq(getMembershipInstancesBody).SetReadTimeout(SimpleOperationTimeout)
 	if err := conn.ExecTyped(req, &membershipInstancesSlice); err != nil {
@@ -57,47 +69,3 @@ func getMembershipInstancesFromConn(conn *connector.Conn) (*MembershipInstances,
 
 	return &membershipInstances, nil
 }
-
-var (
-	probeInstancesBody = `
-local cartridge = require('cartridge')
-
-local uris = ...
-
-for _, uri in ipairs(uris) do
-    local ok, err = cartridge.admin_probe_server(uri)
-    assert(ok, err)
-end
-`
-
-	getMembershipInstancesBody = `
-local membership = require('membership')
-
-local instances = {}
-
-local members = membership.members()
-
-for uri, member in pairs(members) do
-	local uuid
-	if member.payload ~= nil and member.payload.uuid ~= nil then
-		uuid = member.payload.uuid
-	end
-
-	local alias
-	if member.payload ~= nil and member.payload.alias ~= nil then
-		alias = member.payload.alias
-	end
-
-	local instance = {
-		uri = uri,
-		alias = alias,
-		uuid = uuid,
-		status = member.status,
-	}
-
-	table.insert(instances, instance)
-end
-
-return unpack(instances)
-`
-)
