+++ cli/replicasets/templates/cluster_is_healthy_body.lua
@@ -1,0 +1,3 @@
+
+local cartridge = require('cartridge')
+return cartridge.is_healthy()
