+++ cli/replicasets/templates/edit_instance_body.lua
@@ -1,0 +1,10 @@
+
+local cartridge = require('cartridge')
+
+local servers = ...
+
+local res, err = cartridge.admin_edit_topology({
+	servers = servers,
+})
+
+assert(err == nil, tostring(err))
