+++ cli/replicasets/templates/probe_instances_body.lua
@@ -1,0 +1,9 @@
+
+local cartridge = require('cartridge')
+
+local uris = ...
+
+for _, uri in ipairs(uris) do
+    local ok, err = cartridge.admin_probe_server(uri)
+    assert(ok, err)
+end
