+++ cli/replicasets/templates/table.lua
@@ -1,0 +1,1 @@
+{ {{ .OptsString }} }
