+++ cli/replicasets/topology_replicasets.go
@@ -3,6 +3,7 @@ package replicasets
 import (
 	"fmt"
 
+	"github.com/tarantool/cartridge-cli/cli/codegen/static"
 	"github.com/tarantool/cartridge-cli/cli/common"
 	"github.com/tarantool/cartridge-cli/cli/connector"
 	"github.com/tarantool/cartridge-cli/cli/templates"
@@ -49,6 +50,12 @@ var (
 func init() {
 	var err error
 
+	formatTopologyReplicasetFuncTemplate, err := static.GetStaticFileContent(ReplicasetsLuaTemplateFS,
+		"format_topology_replicaset_func.lua")
+	if err != nil {
+		panic(fmt.Errorf("Failed to get static file content: %s", err))
+	}
+
 	formatTopologyReplicasetFunc, err := templates.GetTemplatedStr(
 		&formatTopologyReplicasetFuncTemplate, map[string]string{
 			"FormatTopologyReplicasetFuncName": formatTopologyReplicasetFuncName,
@@ -57,6 +64,12 @@ func init() {
 
 	if err != nil {
 		panic(fmt.Errorf("Failed to compute get topology replica set function body: %s", err))
+	}
+
+	getTopologyReplicasetsBodyTemplate, err := static.GetStaticFileContent(ReplicasetsLuaTemplateFS,
+		"topology_replicasets_body.lua")
+	if err != nil {
+		panic(fmt.Errorf("Failed to get static file content: %s", err))
 	}
 
 	getTopologyReplicasetsBody, err = templates.GetTemplatedStr(
@@ -125,58 +138,4 @@ func getTopologyReplicaset(conn *connector.Conn, replicasetAlias string) (*Topol
 
 const (
 	formatTopologyReplicasetFuncName = "format_topology_replicaset"
-)
-
-var (
-	getTopologyReplicasetsBodyTemplate = `
-local cartridge = require('cartridge')
-
-{{ .FormatTopologyReplicasetFunc }}
-
-local topology_replicasets = {}
-
-local replicasets, err = cartridge.admin_get_replicasets()
-assert(err == nil, tostring(err))
-
-for _, replicaset in pairs(replicasets) do
-	local topology_replicaset = {{ .FormatTopologyReplicasetFuncName }}(replicaset)
-	table.insert(topology_replicasets, topology_replicaset)
-end
-
-return unpack(topology_replicasets)
-`
-
-	formatTopologyReplicasetFuncTemplate = `
-local function {{ .FormatTopologyReplicasetFuncName }}(replicaset)
-	local instances = {}
-	for _, server in pairs(replicaset.servers) do
-		local instance = {
-			alias = server.alias,
-			uuid = server.uuid,
-			uri = server.uri,
-			zone = server.zone,
-		}
-		table.insert(instances, instance)
-	end
-
-	local leader_uuid
-	if replicaset.active_master ~= nil then
-		leader_uuid = replicaset.active_master.uuid
-	end
-
-	local topology_replicaset = {
-		uuid = replicaset.uuid,
-		alias = replicaset.alias,
-		status = replicaset.status,
-		roles = replicaset.roles,
-		all_rw = replicaset.all_rw,
-		weight = replicaset.weight,
-		vshard_group = replicaset.vshard_group,
-		instances = instances,
-		leader_uuid = leader_uuid,
-	}
-
-	return topology_replicaset
-end
-`
 )
