+++ cli/replicasets/vshard_group.go
@@ -3,6 +3,7 @@ package replicasets
 import (
 	"fmt"
 
+	"github.com/tarantool/cartridge-cli/cli/codegen/static"
 	"github.com/tarantool/cartridge-cli/cli/connector"
 
 	"github.com/apex/log"
@@ -13,6 +14,12 @@ func ListVshardGroups(ctx *context.Ctx, args []string) error {
 	conn, err := connectToSomeRunningInstance(ctx)
 	if err != nil {
 		return fmt.Errorf("Failed to connect to Tarantool instance: %s", err)
+	}
+
+	getKnownVshardGroupsBody, err := static.GetStaticFileContent(ReplicasetsLuaTemplateFS,
+		"known_vshard_groups_body.lua")
+	if err != nil {
+		return fmt.Errorf("Failed to get static file content: %s", err)
 	}
 
 	req := connector.EvalReq(getKnownVshardGroupsBody)
@@ -36,18 +43,3 @@ func ListVshardGroups(ctx *context.Ctx, args []string) error {
 
 	return nil
 }
-
-var (
-	getKnownVshardGroupsBody = `
-local vshard_utils = require('cartridge.vshard-utils')
-
-local known_groups = vshard_utils.get_known_groups()
-
-local known_groups_names = {}
-for group_name in pairs(known_groups) do
-	table.insert(known_groups_names, group_name)
-end
-
-return unpack(known_groups_names)
-`
-)
